hypervisor: arm: lib.c: change type of mpidr from u32 to unsigned long
Unsigned long is 32 bits on AArch32 and 64 bits on AArch64. This
allows us to reuse the file on AArch64.

+++ hypervisor/arch/arm/lib.c
@@ -17,7 +17,7 @@
 
 int phys_processor_id(void)
 {
-	u32 mpidr;
+	unsigned long mpidr;
 
 	arm_read_sysreg(MPIDR_EL1, mpidr);
 	return mpidr & MPIDR_CPUID_MASK;
